framework/script/requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
clang==17.0.6
click==8.1.7
cryptography==42.0.7
cxxfilt==0.3.0
cycler==0.12.1
Deprecated==1.2.14
et-xmlfile==1.1.0
exceptiongroup==1.2.1
feather-format==0.4.1
fonttools==4.53.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpath-ng==1.6.1
kiwisolver==1.4.5
libclang==18.1.1
lizard==1.17.10
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.24.4
openpyxl==3.1.3
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
ply==3.11
psycopg2-binary==2.9.9
pyarrow==16.1.0
pycparser==2.22
PyDriller==2.6
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sarif-tools==2.0.0
scikit-learn==1.3.2
# scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.67.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
types-pytz==2024.1.0.20240417
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
whatthepatch==1.0.5
wrapt==1.16.0
zipp==3.19.1
